Fixed warning conversion double to int
git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@2700 b38c6c8f-bee8-465d-8448-9abe6915d496

+++ top/geometry/include/TOPGeometryPar.h
@@ -228,12 +228,12 @@ namespace Belle2 {
       /** Get pad size in x direction
        * @return pad size in x
        */
-      int padx() const {return m_padx; }
+      double padx() const {return m_padx; }
 
       /** Get pad size in y direction
        * @return pad size in x
        */
-      int pady() const {return m_pady; }
+      double pady() const {return m_pady; }
 
       /** Get the thickness of the glue between the PMT stack and the wedge
        * @return the thickness of the glue between the PMT stack and the wedge
